deleted watch-utils and added the utilities in env.ts
