+++ .changeset/nasty-donkeys-collect.md
@@ -1,0 +1,5 @@
+---
+'@eth-optimism/integration-tests': patch
+---
+
+Deletes watcher-utils.ts. Moves it's utilities into env.ts.
